ATT 17/05

+++ TCCMadeireira/Models/ProdVenda.cs
@@ -9,15 +9,15 @@ namespace TCCMadeireira.Models
     class ProdVenda
     {
         private Produto produto;
-        private int quantidade;
+        private decimal quantidade;
 
-        public ProdVenda(Produto produto, int quantidade)
+        public ProdVenda(Produto produto, decimal quantidade)
         {
             this.Produto = produto;
             this.Quantidade = quantidade;
         }
 
-        public int Quantidade { get => quantidade; set => quantidade = value; }
+        public decimal Quantidade { get => quantidade; set => quantidade = value; }
         internal Produto Produto { get => produto; set => produto = value; }
     }
 }
