+++ TCCMadeireira/Models/Produto.cs
@@ -12,7 +12,7 @@ namespace TCCMadeireira.Models
         private string nome;
         private Fornecedor fornecedor;
         private int valor;
-        private int quantidade;
+        private decimal quantidade;
         private string obs;
         private DateTime dataInfo;
 
@@ -20,11 +20,11 @@ namespace TCCMadeireira.Models
         public string Nome { get => nome; set => nome = value; }
         internal Fornecedor Fornecedor { get => fornecedor; set => fornecedor = value; }
         public int Valor { get => valor; set => valor = value; }
-        public int Quantidade { get => quantidade; set => quantidade = value; }
+        public decimal Quantidade { get => quantidade; set => quantidade = value; }
         public DateTime DataInfo { get => dataInfo; set => dataInfo = value; }
         public string Obs { get => obs; set => obs = value; }
 
-        public Produto(int id, string nome, Fornecedor fornecedor, int valor, int quantidade, DateTime dataInfo, string obs) {
+        public Produto(int id, string nome, Fornecedor fornecedor, int valor, decimal quantidade, DateTime dataInfo, string obs) {
             this.Id = id;
             this.Nome = nome.ToUpper().Trim();
             this.Fornecedor = fornecedor;
@@ -33,7 +33,7 @@ namespace TCCMadeireira.Models
             this.Quantidade = quantidade;
             this.DataInfo = dataInfo;
         }
-        public Produto(string nome, Fornecedor fornecedor, int valor, int quantidade, DateTime dataInfo, string obs)
+        public Produto(string nome, Fornecedor fornecedor, int valor, decimal quantidade, DateTime dataInfo, string obs)
         {
             this.Nome = nome.ToUpper().Trim();
             this.obs = obs;
